publish new composite*.xml files to correct folder on download.jb.org
git-svn-id: http://svn.jboss.org/repos/jbosstools/trunk@43711 a97e2381-89e5-4abb-bab3-167db6db766c

+++ build/util/cleanup/jbosstools-cleanup.sh
@@ -89,11 +89,15 @@ clean ()
 		# generate metadata in the nightly/core/trunk/ folder to composite the remaining sites into one
 		getSubDirs $sd 1; #return #getSubDirsReturn
 		getSubDirsCount $getSubDirsReturn; #return $getSubDirsCountReturn
-		mkdir -p /tmp/cleanup-fresh-metadata/
-		regenCompositeMetadata "$getSubDirsReturn" "$getSubDirsCountReturn" "org.eclipse.equinox.internal.p2.metadata.repository.CompositeMetadataRepository" "/tmp/cleanup-fresh-metadata/compositeContent.xml"
-		regenCompositeMetadata "$getSubDirsReturn" "$getSubDirsCountReturn" "org.eclipse.equinox.internal.p2.artifact.repository.CompositeArtifactRepository" "/tmp/cleanup-fresh-metadata/compositeArtifacts.xml"
-		rsync --rsh=ssh --protocol=28 -q /tmp/cleanup-fresh-metadata/composite*.xml tools@filemgmt.jboss.org:$sd/
-		rm -fr /tmp/cleanup-fresh-metadata/
+		if [[ $getSubDirsCountReturn -gt 0 ]]; 
+			mkdir -p /tmp/cleanup-fresh-metadata/
+			regenCompositeMetadata "$getSubDirsReturn" "$getSubDirsCountReturn" "org.eclipse.equinox.internal.p2.metadata.repository.CompositeMetadataRepository" "/tmp/cleanup-fresh-metadata/compositeContent.xml"
+			regenCompositeMetadata "$getSubDirsReturn" "$getSubDirsCountReturn" "org.eclipse.equinox.internal.p2.artifact.repository.CompositeArtifactRepository" "/tmp/cleanup-fresh-metadata/compositeArtifacts.xml"
+			rsync --rsh=ssh --protocol=28 -q /tmp/cleanup-fresh-metadata/composite*.xml tools@filemgmt.jboss.org:$sd/
+			rm -fr /tmp/cleanup-fresh-metadata/
+		else
+			# TODO delete composite*.xml from $sd/ folder if there are no subdirs present
+		fi
 	done
 	echo "" | tee -a $log	
 }
